Move catalog search classes to Common directory
Re #23027
