+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/CatalogSearcher.cpp
