+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/ISearcher.h
