+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/SearchModel.cpp
