+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/SearchModel.h
