+++ qt/scientific_interfaces/ISISReflectometry/GUI/Common/SearchResult.h
