+++ qt/scientific_interfaces/test/ISISReflectometry/ReflMockObjects.h
@@ -11,6 +11,8 @@
 #include "GUI/Batch/IBatchJobRunner.h"
 #include "GUI/Batch/IBatchPresenter.h"
 #include "GUI/Common/IMessageHandler.h"
+#include "GUI/Common/ISearcher.h"
+#include "GUI/Common/SearchModel.h"
 #include "GUI/Event/IEventPresenter.h"
 #include "GUI/Experiment/IExperimentPresenter.h"
 #include "GUI/Instrument/IInstrumentPresenter.h"
@@ -20,8 +22,6 @@
 #include "GUI/Runs/IAutoreduction.h"
 #include "GUI/Runs/IRunNotifier.h"
 #include "GUI/Runs/IRunsPresenter.h"
-#include "GUI/Runs/ISearcher.h"
-#include "GUI/Runs/SearchModel.h"
 #include "GUI/Save/IAsciiSaver.h"
 #include "GUI/Save/ISavePresenter.h"
 #include "MantidAPI/AlgorithmManager.h"
